requirements.txt
streamlit
langchain[community]
langchain_google_genai
googlesearch-python
requests
readability-lxml
beautifulsoup4